Switch to use transient
`transient` is a replacement to `magit-popup` and even `magit` has already
switched to use `transient`.

The switch is pretty idiomatic, however there are a few differences.

The default layout is slightly different but I didn't try to replace it as the
new one seems to match what `magit` is presenting these days.

The concept of `:default-action` seems to be gone, so it is not replaced.

The `magit-with-pre-popup-buffer` doesn't seem to have an obvious replacement,
however the documentation for `transient--original-buffer` suggest that the
`transient` is not changing the current buffer.