Declare python-pytest--current-defun obsolete
The new functions that should be used python-pytest--path-def-at-point
are python-pytest--path-class-at-point.

python-pytest--current-defun could obtain the identifiers for
functions and classes. If the current point was inside a function,
there was no way to get the test id of the current class. By using
python-pytest--path-def-at-point and
python-pytest--path-class-at-point, there is no such limitation.