rename function python-pytest--current-defun for better readibility
In a recent previous commit, we have added 2 functions:
python-pytest--node-id-def-at-point-treesit and
python-pytest--node-id-class-at-point-treesit. These two function
names clearly explain that they either get the node id of the def at
or class at point.

The name python-pytest--current-defun can be confusing for some
readers because the reader might think that it can only get the def at
point, not the class at point. The new name
python-pytest--node-id-def-or-class-at-point makes it clear that it
can get both the def or class at point.